pyproject.toml
[project]
name = "document-refinery"
version = "0.1.0"
description = "Production document refinery with triage and multi-strategy extraction"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    # Core dependencies
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "numpy>=1.21.0",
    "loguru>=0.7.3",
    
    # Document processing
    "pymupdf>=1.23.0",  # fitz for PDF analysis
    "pdfplumber>=0.9.0",  # Fast text extraction
    "pillow>=9.0.0",  # Image processing
    
    # OCR and Vision
    "pytesseract>=0.3.10",  # Tesseract OCR
    "pdf2image>=1.16.0",  # PDF to image conversion
    
    # Layout extraction
    "docling>=2.76.0",  # Advanced layout analysis
    
    # Optional: VLM integration
    "tiktoken>=0.12.0",  # Token counting
    "requests>=2.28.0",  # HTTP requests
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "flake8>=5.0.0",
    "mypy>=1.0.0",
    "pre-commit>=2.20.0",
]

vision = [
    "torch>=2.0.0",
    "transformers>=4.30.0",
    "opencv-python>=4.8.0",
    "openai>=1.0.0",
]

all = [
    "document-refinery[dev,vision]"
]

[project.urls]
Homepage = "https://github.com/your-org/document-refinery"
Repository = "https://github.com/your-org/document-refinery.git"
Issues = "https://github.com/your-org/document-refinery/issues"

[project.scripts]
refinery = "refinery:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = [
    "tests",
]

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/test_*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
]
